+++ src/conduits/java/com/enderio/conduits/common/types/ItemClientConduitData.java
@@ -4,6 +4,7 @@ import com.enderio.EnderIO;
 import com.enderio.api.conduit.IClientConduitData;
 import com.enderio.api.conduit.IConduitType;
 import com.enderio.api.misc.Vector2i;
+import com.enderio.base.common.lang.EIOLang;
 import com.enderio.conduits.common.init.EnderConduitTypes;
 import com.enderio.conduits.common.network.C2SSetConduitExtendedData;
 import com.enderio.core.client.gui.widgets.CheckBox;
@@ -23,14 +24,18 @@ public class ItemClientConduitData implements IClientConduitData<ItemExtendedDat
     public List<AbstractWidget> createWidgets(Screen screen, Supplier<BlockPos> blockPos, Supplier<IConduitType<?>> type, ItemExtendedData extendedConduitData, Supplier<Direction> direction, Vector2i widgetsStart) {
         // TODO: Method of doing sync that does not require CoreNetwork in API.
         List<AbstractWidget> widgets = new ArrayList<>();
-        widgets.add(new CheckBox(EnderIO.loc("textures/gui/round_robin.png"), widgetsStart.add(110, 20), () -> extendedConduitData.get(direction.get()).roundRobin, bool -> {
-            extendedConduitData.compute(direction.get()).roundRobin = bool;
-            CoreNetwork.sendToServer(new C2SSetConduitExtendedData(blockPos.get(), type.get(), extendedConduitData));
-        }));
-        widgets.add(new CheckBox(EnderIO.loc("textures/gui/self_feed.png"), widgetsStart.add(130, 20), () -> extendedConduitData.get(direction.get()).selfFeed, bool -> {
-            extendedConduitData.compute(direction.get()).selfFeed = bool;
-            CoreNetwork.sendToServer(new C2SSetConduitExtendedData(blockPos.get(), type.get(), extendedConduitData));
-        }));
+        widgets.add(new CheckBox(EnderIO.loc("textures/gui/round_robin.png"), widgetsStart.add(110, 20),
+            () -> extendedConduitData.get(direction.get()).roundRobin,
+            bool -> {
+                extendedConduitData.compute(direction.get()).roundRobin = bool;
+                CoreNetwork.sendToServer(new C2SSetConduitExtendedData(blockPos.get(), type.get(), extendedConduitData));
+            }, () -> EIOLang.ROUND_ROBIN_ENABLED, () -> EIOLang.ROUND_ROBIN_DISABLED));
+        widgets.add(new CheckBox(EnderIO.loc("textures/gui/self_feed.png"), widgetsStart.add(130, 20),
+            () -> extendedConduitData.get(direction.get()).selfFeed,
+            bool -> {
+                extendedConduitData.compute(direction.get()).selfFeed = bool;
+                CoreNetwork.sendToServer(new C2SSetConduitExtendedData(blockPos.get(), type.get(), extendedConduitData));
+            }, () -> EIOLang.SELF_FEED_ENABLED, () -> EIOLang.SELF_FEED_DISABLED));
         return widgets;
     }