+++ src/core/java/com/enderio/core/client/gui/widgets/CheckBox.java
@@ -3,8 +3,10 @@ package com.enderio.core.client.gui.widgets;
 import com.enderio.api.misc.Vector2i;
 import com.enderio.core.EnderCore;
 import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractButton;
+import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.client.gui.narration.NarrationElementOutput;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
@@ -16,18 +18,27 @@ public class CheckBox extends AbstractButton {
 
     private final Supplier<Boolean> getter;
     private final Consumer<Boolean> setter;
+    private final Supplier<Component> enabledTooltip;
+    private final Supplier<Component> disabledTooltip;
+
     private final ResourceLocation texture;
     private static final ResourceLocation TEXTURE = EnderCore.loc("textures/gui/checkbox.png");
 
     public CheckBox(Vector2i pos, Supplier<Boolean> getter, Consumer<Boolean> setter) {
-        this(TEXTURE, pos, getter, setter);
+        this(TEXTURE, pos, getter, setter, () -> null, () -> null);
     }
 
-    public CheckBox(ResourceLocation texture, Vector2i pos, Supplier<Boolean> getter, Consumer<Boolean> setter) {
+    public CheckBox(Vector2i pos, Supplier<Boolean> getter, Consumer<Boolean> setter, Supplier<Component> enabledTooltip, Supplier<Component> disabledTooltip) {
+        this(TEXTURE, pos, getter, setter, enabledTooltip, disabledTooltip);
+    }
+
+    public CheckBox(ResourceLocation texture, Vector2i pos, Supplier<Boolean> getter, Consumer<Boolean> setter, Supplier<Component> enabledTooltip, Supplier<Component> disabledTooltip) {
         super(pos.x(), pos.y(), 14, 14, Component.empty());
         this.getter = getter;
         this.setter = setter;
         this.texture = texture;
+        this.enabledTooltip = enabledTooltip;
+        this.disabledTooltip = disabledTooltip;
     }
 
     @Override
@@ -51,6 +62,10 @@ public class CheckBox extends AbstractButton {
 
         RenderSystem.disableBlend();
         RenderSystem.disableDepthTest();
+
+        if (this.isHovered && enabledTooltip.get() != null && disabledTooltip.get() != null) {
+            setTooltip(Tooltip.create((getter.get() ? enabledTooltip : disabledTooltip).get().copy().withStyle(ChatFormatting.WHITE)));
+        }
     }
 
     @Override