+++ CHANGELOG.md
@@ -29,11 +29,13 @@
 ### Added
 - Added missing tooltips for IO Config button and Neighbour button.
 - Better multismelting support for the alloy smelter
+- Added tooltips to Conduit round-robin and self feed buttons
 
 ### Changed
 - Grindstone crafting for grains of infinity and powdered coal now uses an in-world craft by right-clicking with flint in the off-hand and the ingredient in the main-hand. Also works with obsidian and crying obsidian (the latter of which has a chance to give a better rate of return on the craft).
 - Lowered minimum Forge version to 47.0.42.
 - Conduits are now placed in your inventory when shift right clicked with the Yeta Wrench
+- Gave CheckBox widgets the ability to have tooltips
 
 ### Removed
 - Removed the grave system